Conflicts:
	cs_interface/cpp/Accelerometer.cpp
	cs_interface/cpp/Accelerometer.hpp
	cs_interface/cpp/Gyroscope.hpp
	cs_interface/cpp/Magnetometer.hpp
	cs_interface/cpp/sensor.cpp

+++ cs_interface/cpp/Accelerometer.cpp
@@ -8,7 +8,7 @@ Accelerometer::Accelerometer()
     gain[2] = 0.00349265;
 }
 
-void Accelerometer::updateAccel(int ax, int ay, int az)
+void Accelerometer::updateAccel(int16_t ax, int16_t ay, int16_t az)
 {
     x_raw = ax;
     y_raw = ay;
