+++ cs_interface/cpp/Gyroscope.cpp
@@ -38,7 +38,7 @@ void Gyroscope::zeroCalibrate(unsigned int totSamples, unsigned int sampleDelayM
 	offset[2] = -tmpOffset[2] / totSamples;
 }
 
-void Gyroscope::updateGyro(int gyroX, int gyroY, int gyroZ)
+void Gyroscope::updateGyro(int16_t gyroX, int16_t gyroY, int16_t gyroZ)
 {
 	raw_x = gyroX;
 	x = -gyroX;
